requirements.txt
Flask
Flask-Migrate
flask-migrate
flask_jwt_extended
flask_marshmallow
flask_restful
flask_sqlalchemy
passlib
python-dotenv